增加角色表的CRUD;

+++ src/logical/role/role.service.ts
@@ -11,48 +11,31 @@ export class RoleService {
   ) {}
 
   async create(createRoleDto: RoleItemDto) {
-    try {
-      createRoleDto.roleValue = `ROLE_${createRoleDto.roleName}`
-      return await this.roleRepository.save(createRoleDto)
-    } catch (e) {
-      console.log(e)
-      return void 0
-    }
+    createRoleDto.roleValue = `ROLE_${createRoleDto.roleName}`
+    return await this.roleRepository.save(createRoleDto)
   }
 
   async findAll() {
-    try {
-      return await this.roleRepository.find()
-    } catch (e) {
-      console.log(e)
-      return void 0
-    }
+    return await this.roleRepository.find()
   }
 
   async findOne(id: number) {
-    try {
-      return await this.roleRepository.findOne({ id })
-    } catch (e) {
-      console.log(e)
-      return void 0
-    }
+    return await this.roleRepository.findOne({ id })
   }
 
-  async update(id: number, updateRoleDto: UpdateRoleDto) {
-    try {
-      return await this.roleRepository.update(id, updateRoleDto)
-    } catch (e) {
-      console.log(e)
-      return void 0
+  async update(id: number, updateRoleDto: RoleItemDto) {
+    if (updateRoleDto.roleName) {
+      updateRoleDto.roleValue = `ROLE_${updateRoleDto.roleName}`
     }
+    return await this.roleRepository.update(id, updateRoleDto)
   }
 
   async remove(id: number) {
-    try {
+    const role = await this.roleRepository.findOne(id)
+    if (role) {
       return await this.roleRepository.delete(id)
-    } catch (e) {
-      console.log(e)
-      return void 0
+    } else {
+      return '没有该角色'
     }
   }
 }
